refactor(codegen): split EVM IR modules
